Refactor libndt to be a single-header library
