+++ .clang-format
@@ -1,1 +1,4 @@
 BasedOnStyle: Google
+ColumnLimit: 0
+AllowShortIfStatementsOnASingleLine: true
+AllowShortCaseLabelsOnASingleLine: true
