+++ tests_main.cpp
@@ -5,3 +5,6 @@
 // See https://github.com/catchorg/Catch2/blob/master/docs/slow-compiles.md#top
 #define CATCH_CONFIG_MAIN
 #include "catch.hpp"
+
+#include "json.hpp"
+#include "libndt.hpp"
